cluster

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: nakkim <nakkim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 21:03:53 by nakkim            #+#    #+#             */
-/*   Updated: 2021/11/19 13:23:11 by nakkim           ###   ########.fr       */
+/*   Updated: 2021/11/21 15:47:38 by nakkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	ft_atoi(const char *ptr)
 {
-	int			sign;
-	long long	num;
+	int	sign;
+	int	num;
 
 	sign = 1;
 	num = 0;
